Add files via upload
The initial script was modified to be able to load and clean more docx files,  and write the output to one single docx file. I.e. if you have a book divided into more docx files, it will clean them all, and write the output sorted, from the oldest file to the newest, into a new single file, and the pages of the book remain sorted ascending.